Refactor app into modular pages
