+++ README.md
@@ -9,6 +9,7 @@
 - **ลบโฟลเดอร์เก่า** และ **Clear Cache** อัตโนมัติเมื่อเปลี่ยน URL
 - **เก็บ Repository ใน Web Cache** แทนการบันทึกลงเครื่อง
 - **สุ่มเลขบัตรประชาชนไทย** พร้อมดูข้อมูลจังหวัดและภาค
+- **หน้า Home** สำหรับเลือกใช้งานแต่ละฟีเจอร์
 
 ## วิธีใช้งาน (Local Machine)
