+++ app.py
@@ -1,306 +1,17 @@
 import streamlit as st
-import os
-import re
-import random
+from features import git_compare_page, thai_id_page
 
-from git_compare import repo_manager, diff_utils
+st.set_page_config(page_title="Multi Feature App", layout="wide")
 
-# ฟังก์ชันโหลด CSS จากไฟล์
-def load_custom_css(css_path: str) -> str:
-    """โหลดไฟล์ CSS มาเป็น string"""
-    if os.path.exists(css_path):
-        with open(css_path, "r", encoding="utf-8") as f:
-            return f"<style>{f.read()}</style>"
-    return ""
-
-# ----- ตั้งค่าหน้าตา Streamlit ----- #
-st.set_page_config(
-    page_title="Git Compare - Regex File Mapping",
-    layout="wide"
+page = st.sidebar.selectbox(
+    "เลือกฟีเจอร์",
+    ("Home", "Git Compare", "Thai ID Generator"),
 )
 
-tab_git, tab_id = st.tabs(["Git Compare", "Thai ID Generator"])
-
-# ===== ข้อมูลจังหวัดและภาคจากเลขบัตรประชาชน ===== #
-province_region_map = {
-    "10": ("กรุงเทพมหานคร", "ภาคกลาง"),
-    "11": ("สมุทรปราการ", "ภาคกลาง"),
-    "12": ("นนทบุรี", "ภาคกลาง"),
-    "13": ("ปทุมธานี", "ภาคกลาง"),
-    "14": ("พระนครศรีอยุธยา", "ภาคกลาง"),
-    "15": ("อ่างทอง", "ภาคกลาง"),
-    "16": ("ลพบุรี", "ภาคกลาง"),
-    "17": ("สิงห์บุรี", "ภาคกลาง"),
-    "18": ("ชัยนาท", "ภาคกลาง"),
-    "19": ("สระบุรี", "ภาคกลาง"),
-    "20": ("ชลบุรี", "ภาคตะวันออก"),
-    "21": ("ระยอง", "ภาคตะวันออก"),
-    "22": ("จันทบุรี", "ภาคตะวันออก"),
-    "23": ("ตราด", "ภาคตะวันออก"),
-    "24": ("ฉะเชิงเทรา", "ภาคตะวันออก"),
-    "25": ("ปราจีนบุรี", "ภาคตะวันออก"),
-    "26": ("นครนายก", "ภาคกลาง"),
-    "27": ("สระแก้ว", "ภาคตะวันออก"),
-    "30": ("นครราชสีมา", "ภาคตะวันออกเฉียงเหนือ"),
-    "31": ("บุรีรัมย์", "ภาคตะวันออกเฉียงเหนือ"),
-    "32": ("สุรินทร์", "ภาคตะวันออกเฉียงเหนือ"),
-    "33": ("ศรีสะเกษ", "ภาคตะวันออกเฉียงเหนือ"),
-    "34": ("อุบลราชธานี", "ภาคตะวันออกเฉียงเหนือ"),
-    "35": ("ยโสธร", "ภาคตะวันออกเฉียงเหนือ"),
-    "36": ("ชัยภูมิ", "ภาคตะวันออกเฉียงเหนือ"),
-    "37": ("อำนาจเจริญ", "ภาคตะวันออกเฉียงเหนือ"),
-    "38": ("บึงกาฬ", "ภาคตะวันออกเฉียงเหนือ"),
-    "39": ("หนองบัวลำภู", "ภาคตะวันออกเฉียงเหนือ"),
-    "40": ("ขอนแก่น", "ภาคตะวันออกเฉียงเหนือ"),
-    "41": ("อุดรธานี", "ภาคตะวันออกเฉียงเหนือ"),
-    "42": ("เลย", "ภาคตะวันออกเฉียงเหนือ"),
-    "43": ("หนองคาย", "ภาคตะวันออกเฉียงเหนือ"),
-    "44": ("มหาสารคาม", "ภาคตะวันออกเฉียงเหนือ"),
-    "45": ("ร้อยเอ็ด", "ภาคตะวันออกเฉียงเหนือ"),
-    "46": ("กาฬสินธุ์", "ภาคตะวันออกเฉียงเหนือ"),
-    "47": ("สกลนคร", "ภาคตะวันออกเฉียงเหนือ"),
-    "48": ("นครพนม", "ภาคตะวันออกเฉียงเหนือ"),
-    "49": ("มุกดาหาร", "ภาคตะวันออกเฉียงเหนือ"),
-    "50": ("เชียงใหม่", "ภาคเหนือ"),
-    "51": ("ลำพูน", "ภาคเหนือ"),
-    "52": ("ลำปาง", "ภาคเหนือ"),
-    "53": ("อุตรดิตถ์", "ภาคเหนือ"),
-    "54": ("แพร่", "ภาคเหนือ"),
-    "55": ("น่าน", "ภาคเหนือ"),
-    "56": ("พะเยา", "ภาคเหนือ"),
-    "57": ("เชียงราย", "ภาคเหนือ"),
-    "58": ("แม่ฮ่องสอน", "ภาคเหนือ"),
-    "60": ("นครสวรรค์", "ภาคเหนือ"),
-    "61": ("อุทัยธานี", "ภาคเหนือ"),
-    "62": ("กำแพงเพชร", "ภาคเหนือ"),
-    "63": ("ตาก", "ภาคเหนือ"),
-    "64": ("สุโขทัย", "ภาคเหนือ"),
-    "65": ("พิษณุโลก", "ภาคเหนือ"),
-    "66": ("พิจิตร", "ภาคเหนือ"),
-    "67": ("เพชรบูรณ์", "ภาคเหนือ"),
-    "70": ("ราชบุรี", "ภาคตะวันตก"),
-    "71": ("กาญจนบุรี", "ภาคตะวันตก"),
-    "72": ("สุพรรณบุรี", "ภาคกลาง"),
-    "73": ("นครปฐม", "ภาคกลาง"),
-    "74": ("สมุทรสาคร", "ภาคกลาง"),
-    "75": ("สมุทรสงคราม", "ภาคกลาง"),
-    "76": ("เพชรบุรี", "ภาคตะวันตก"),
-    "77": ("ประจวบคีรีขันธ์", "ภาคตะวันตก"),
-    "80": ("นครศรีธรรมราช", "ภาคใต้"),
-    "81": ("กระบี่", "ภาคใต้"),
-    "82": ("พังงา", "ภาคใต้"),
-    "83": ("ภูเก็ต", "ภาคใต้"),
-    "84": ("สุราษฎร์ธานี", "ภาคใต้"),
-    "85": ("ระนอง", "ภาคใต้"),
-    "86": ("ชุมพร", "ภาคใต้"),
-    "90": ("สงขลา", "ภาคใต้"),
-    "91": ("สตูล", "ภาคใต้"),
-    "92": ("ตรัง", "ภาคใต้"),
-    "93": ("พัทลุง", "ภาคใต้"),
-    "94": ("ปัตตานี", "ภาคใต้"),
-    "95": ("ยะลา", "ภาคใต้"),
-    "96": ("นราธิวาส", "ภาคใต้"),
-}
-
-def generate_thai_id() -> str:
-    """สุ่มเลขบัตรประชาชนไทย 13 หลัก"""
-    digits = [random.randint(0, 9) for _ in range(12)]
-    checksum = sum(d * (13 - i) for i, d in enumerate(digits, start=1))
-    check_digit = (11 - (checksum % 11)) % 10
-    digits.append(check_digit)
-    return "".join(str(d) for d in digits)
-
-
-def get_province_info(citizen_id: str):
-    """ดึงข้อมูลจังหวัดและภาคจากเลขบัตร"""
-    code = citizen_id[1:3]
-    return province_region_map.get(code, ("ไม่ทราบ", "ไม่ทราบ"))
-
-def git_compare_tab():
-    st.title("🔍 Git Repository Compare - Regex File Mapping, Dark Theme")
-
-    # ========== (1) ตรวจจับ URL เปลี่ยน + ลบโฟลเดอร์เดิม + Clear cache ========== #
-    if "old_repo_url" not in st.session_state:
-        st.session_state.old_repo_url = ""
-    
-    repo_url = st.text_input("📂 Git Repository URL:", "")
-    if "repo_dir" not in st.session_state:
-        st.session_state.repo_dir = ""
-    
-    # ถ้า URL เปลี่ยน => ลบโฟลเดอร์ + clear cache
-    if repo_url != st.session_state.old_repo_url and st.session_state.old_repo_url != "":
-        st.warning(
-            f"Repo URL changed from {st.session_state.old_repo_url} to {repo_url}. Removing old folder & clearing cache."
-        )
-        repo_manager.remove_dir(st.session_state.repo_dir)
-        repo_manager.clear_repo_cache()
-        st.session_state.repo_dir = ""
-        st.cache_data.clear()
-    
-    st.session_state.old_repo_url = repo_url
-    
-    # ===== ปุ่ม Reload Cache ด้วยตัวเอง ===== #
-    if st.button("Reload Cache"):
-        repo_manager.clear_repo_cache()
-        repo_manager.remove_dir(st.session_state.repo_dir)
-        st.session_state.repo_dir = ""
-        st.cache_data.clear()
-        st.success("Cache cleared manually!")
-    
-    # ===== ปุ่ม Clone ===== #
-    if st.button("Clone Repository"):
-        with st.spinner("Cloning or Checking..."):
-            st.session_state.repo_dir = repo_manager.clone_repo_to_cache(repo_url)
-        st.success("Repository is ready!")
-    
-    # ===== ถ้ามีโฟลเดอร์ => ดึง Branch/Tag ===== #
-    repo_dir = st.session_state.repo_dir
-    if os.path.exists(repo_dir):
-        branches, tags = repo_manager.get_all_branches_and_tags(repo_dir)
-        if not branches and not tags:
-            st.error("No Branches or Tags found in this repository.")
-        else:
-            # ----- Compare Side 1 ----- #
-            st.subheader("Compare Side 1")
-            compare_type_1 = st.radio("Compare type 1", ["Branch", "Tag"], index=0, horizontal=True)
-            ref_list_1 = branches if compare_type_1 == "Branch" else tags
-            side1 = st.selectbox("Select side 1", ref_list_1, key="side1_select") if ref_list_1 else None
-    
-            # ----- Compare Side 2 ----- #
-            st.subheader("Compare Side 2")
-            compare_type_2 = st.radio("Compare type 2", ["Branch", "Tag"], index=0, horizontal=True)
-            ref_list_2 = branches if compare_type_2 == "Branch" else tags
-            side2 = st.selectbox("Select side 2", ref_list_2, key="side2_select") if ref_list_2 else None
-    
-            if side1 and side2:
-                files_side1 = repo_manager.list_files_in_ref(repo_dir, side1)
-                files_side2 = repo_manager.list_files_in_ref(repo_dir, side2)
-    
-                only_in_side1 = files_side1 - files_side2
-                only_in_side2 = files_side2 - files_side1
-                common_files = files_side1.intersection(files_side2)
-    
-                st.subheader("📁 Folder & File Structure Differences")
-                col_left, col_right = st.columns(2)
-                with col_left:
-                    st.write(f"Files only in `{side1}`:")
-                    st.code("\n".join(sorted(only_in_side1)) if only_in_side1 else "— None —")
-                with col_right:
-                    st.write(f"Files only in `{side2}`:")
-                    st.code("\n".join(sorted(only_in_side2)) if only_in_side2 else "— None —")
-    
-                # ----- Custom File Mapping (Manual) ----- #
-                if "file_mapping" not in st.session_state:
-                    st.session_state.file_mapping = {}
-    
-                st.subheader("🔗 Custom File Mapping (Manual)")
-                with st.form("custom_mapping_form", clear_on_submit=True):
-                    f_in_1 = st.text_input(f"File name in {side1}")
-                    f_in_2 = st.text_input(f"File name in {side2}")
-                    submitted_map = st.form_submit_button("Add Mapping")
-                    if submitted_map:
-                        if f_in_1 and f_in_2:
-                            st.session_state.file_mapping[f_in_1] = f_in_2
-                            st.success(f"Added mapping: {f_in_1} => {f_in_2}")
-                        else:
-                            st.warning("Please fill both file names before adding mapping.")
-    
-                # ----- Custom File Mapping (by Regex) ----- #
-                st.subheader("🤖 Regex-based File Mapping")
-                with st.form("regex_mapping_form"):
-                    pattern = st.text_input("Regex Pattern (Side 1)", "")
-                    replacement = st.text_input("Regex Replacement (Side 1)", "")
-                    apply_regex = st.form_submit_button("Apply Regex to Side 1")
-                    if apply_regex:
-                        if pattern:
-                            # ใช้ฟังก์ชัน regex_file_mapping จาก repo_manager
-                            auto_mapping = repo_manager.regex_file_mapping(files_side1, pattern, replacement)
-                            st.session_state.file_mapping.update(auto_mapping)
-                            st.success(f"Applied Regex => {len(auto_mapping)} files mapped.")
-                        else:
-                            st.warning("Please provide a valid regex pattern.")
-    
-                if st.session_state.file_mapping:
-                    st.write("Current Mappings:")
-                    for k, v in st.session_state.file_mapping.items():
-                        st.text(f"{k} => {v}")
-                else:
-                    st.write("No custom file mappings yet.")
-    
-                # ----- รวมไฟล์ที่จะเทียบ (common + mapping) ----- #
-                compare_pairs = {f: f for f in common_files}
-                for old_name, new_name in st.session_state.file_mapping.items():
-                    compare_pairs[old_name] = new_name
-    
-                # ----- Partial Diff: สรุปไฟล์ที่ต่าง ----- #
-                summary_data = []
-                for f1, f2 in compare_pairs.items():
-                    content1 = repo_manager.get_file_content(repo_dir, side1, f1)
-                    content2 = repo_manager.get_file_content(repo_dir, side2, f2)
-                    if content1 and content2 and (content1 != content2):
-                        changed = diff_utils.quick_diff_lines(content1, content2)
-                        if changed > 0:
-                            summary_data.append({
-                                "file1": f1,
-                                "file2": f2,
-                                "changed_lines": changed
-                            })
-    
-                if summary_data:
-                    st.subheader("📝 Differences Summary")
-                    st.table(summary_data)
-    
-                    pair_display_list = [
-                        f"{item['file1']} => {item['file2']} (Changed {item['changed_lines']} lines)"
-                        for item in summary_data
-                    ]
-                    selected_pair_str = st.selectbox("📄 Select File Pair to Compare:", pair_display_list)
-    
-                    reverse_map = {
-                        f"{item['file1']} => {item['file2']} (Changed {item['changed_lines']} lines)": (item['file1'], item['file2'])
-                        for item in summary_data
-                    }
-    
-                    file1, file2 = reverse_map[selected_pair_str]
-                    c1 = repo_manager.get_file_content(repo_dir, side1, file1)
-                    c2 = repo_manager.get_file_content(repo_dir, side2, file2)
-    
-                    st.subheader(f"🔍 Side-by-Side Diff: {file1} => {file2}")
-                    html_diff = diff_utils.make_side_by_side_diff(
-                        c1, c2, f"{side1}:{file1}", f"{side2}:{file2}"
-                    )
-    
-                    # โหลด Custom CSS
-                    css_content = load_custom_css("assets/custom.css")
-                    st.components.v1.html(css_content + html_diff, height=800, scrolling=True)
-                else:
-                    st.success("✅ No differences found among common or mapped files.")
-            else:
-                st.info("Please select both sides (Branch or Tag).")
-    else:
-        st.info("Please clone the repository or provide a valid URL.")
-
-with tab_git:
-    git_compare_tab()
-
-with tab_id:
-    st.header("🎲 สุ่มเลขบัตรประชาชนไทย")
-    count = st.number_input("จำนวนเลขที่ต้องการสุ่ม", min_value=1, max_value=20, value=1, step=1)
-    if st.button("Generate", key="gen_thai_id"):
-        st.session_state.generated_ids = [generate_thai_id() for _ in range(count)]
-        # reset info flags
-        for k in list(st.session_state.keys()):
-            if k.startswith("show_info_"):
-                del st.session_state[k]
-
-    ids = st.session_state.get("generated_ids", [])
-    for idx, cid in enumerate(ids):
-        col1, col2 = st.columns([3,1])
-        col1.write(cid)
-        info_key = f"info_{idx}"
-        if col2.button("Information", key=info_key):
-            st.session_state[f"show_info_{idx}"] = not st.session_state.get(f"show_info_{idx}", False)
-        if st.session_state.get(f"show_info_{idx}", False):
-            prov, region = get_province_info(cid)
-            st.info(f"จังหวัด: {prov} | ภาค: {region}")
-
+if page == "Home":
+    st.title("🛠 Multi Feature App")
+    st.write("เลือกฟีเจอร์จากแถบด้านซ้ายเพื่อเริ่มใช้งาน")
+elif page == "Git Compare":
+    git_compare_page.show()
+elif page == "Thai ID Generator":
+    thai_id_page.show()
