+++ features/thai_id_page.py
@@ -1,0 +1,117 @@
+import random
+import streamlit as st
+
+# ===== ข้อมูลจังหวัดและภาคจากเลขบัตรประชาชน ===== #
+province_region_map = {
+    "10": ("กรุงเทพมหานคร", "ภาคกลาง"),
+    "11": ("สมุทรปราการ", "ภาคกลาง"),
+    "12": ("นนทบุรี", "ภาคกลาง"),
+    "13": ("ปทุมธานี", "ภาคกลาง"),
+    "14": ("พระนครศรีอยุธยา", "ภาคกลาง"),
+    "15": ("อ่างทอง", "ภาคกลาง"),
+    "16": ("ลพบุรี", "ภาคกลาง"),
+    "17": ("สิงห์บุรี", "ภาคกลาง"),
+    "18": ("ชัยนาท", "ภาคกลาง"),
+    "19": ("สระบุรี", "ภาคกลาง"),
+    "20": ("ชลบุรี", "ภาคตะวันออก"),
+    "21": ("ระยอง", "ภาคตะวันออก"),
+    "22": ("จันทบุรี", "ภาคตะวันออก"),
+    "23": ("ตราด", "ภาคตะวันออก"),
+    "24": ("ฉะเชิงเทรา", "ภาคตะวันออก"),
+    "25": ("ปราจีนบุรี", "ภาคตะวันออก"),
+    "26": ("นครนายก", "ภาคกลาง"),
+    "27": ("สระแก้ว", "ภาคตะวันออก"),
+    "30": ("นครราชสีมา", "ภาคตะวันออกเฉียงเหนือ"),
+    "31": ("บุรีรัมย์", "ภาคตะวันออกเฉียงเหนือ"),
+    "32": ("สุรินทร์", "ภาคตะวันออกเฉียงเหนือ"),
+    "33": ("ศรีสะเกษ", "ภาคตะวันออกเฉียงเหนือ"),
+    "34": ("อุบลราชธานี", "ภาคตะวันออกเฉียงเหนือ"),
+    "35": ("ยโสธร", "ภาคตะวันออกเฉียงเหนือ"),
+    "36": ("ชัยภูมิ", "ภาคตะวันออกเฉียงเหนือ"),
+    "37": ("อำนาจเจริญ", "ภาคตะวันออกเฉียงเหนือ"),
+    "38": ("บึงกาฬ", "ภาคตะวันออกเฉียงเหนือ"),
+    "39": ("หนองบัวลำภู", "ภาคตะวันออกเฉียงเหนือ"),
+    "40": ("ขอนแก่น", "ภาคตะวันออกเฉียงเหนือ"),
+    "41": ("อุดรธานี", "ภาคตะวันออกเฉียงเหนือ"),
+    "42": ("เลย", "ภาคตะวันออกเฉียงเหนือ"),
+    "43": ("หนองคาย", "ภาคตะวันออกเฉียงเหนือ"),
+    "44": ("มหาสารคาม", "ภาคตะวันออกเฉียงเหนือ"),
+    "45": ("ร้อยเอ็ด", "ภาคตะวันออกเฉียงเหนือ"),
+    "46": ("กาฬสินธุ์", "ภาคตะวันออกเฉียงเหนือ"),
+    "47": ("สกลนคร", "ภาคตะวันออกเฉียงเหนือ"),
+    "48": ("นครพนม", "ภาคตะวันออกเฉียงเหนือ"),
+    "49": ("มุกดาหาร", "ภาคตะวันออกเฉียงเหนือ"),
+    "50": ("เชียงใหม่", "ภาคเหนือ"),
+    "51": ("ลำพูน", "ภาคเหนือ"),
+    "52": ("ลำปาง", "ภาคเหนือ"),
+    "53": ("อุตรดิตถ์", "ภาคเหนือ"),
+    "54": ("แพร่", "ภาคเหนือ"),
+    "55": ("น่าน", "ภาคเหนือ"),
+    "56": ("พะเยา", "ภาคเหนือ"),
+    "57": ("เชียงราย", "ภาคเหนือ"),
+    "58": ("แม่ฮ่องสอน", "ภาคเหนือ"),
+    "60": ("นครสวรรค์", "ภาคเหนือ"),
+    "61": ("อุทัยธานี", "ภาคเหนือ"),
+    "62": ("กำแพงเพชร", "ภาคเหนือ"),
+    "63": ("ตาก", "ภาคเหนือ"),
+    "64": ("สุโขทัย", "ภาคเหนือ"),
+    "65": ("พิษณุโลก", "ภาคเหนือ"),
+    "66": ("พิจิตร", "ภาคเหนือ"),
+    "67": ("เพชรบูรณ์", "ภาคเหนือ"),
+    "70": ("ราชบุรี", "ภาคตะวันตก"),
+    "71": ("กาญจนบุรี", "ภาคตะวันตก"),
+    "72": ("สุพรรณบุรี", "ภาคกลาง"),
+    "73": ("นครปฐม", "ภาคกลาง"),
+    "74": ("สมุทรสาคร", "ภาคกลาง"),
+    "75": ("สมุทรสงคราม", "ภาคกลาง"),
+    "76": ("เพชรบุรี", "ภาคตะวันตก"),
+    "77": ("ประจวบคีรีขันธ์", "ภาคตะวันตก"),
+    "80": ("นครศรีธรรมราช", "ภาคใต้"),
+    "81": ("กระบี่", "ภาคใต้"),
+    "82": ("พังงา", "ภาคใต้"),
+    "83": ("ภูเก็ต", "ภาคใต้"),
+    "84": ("สุราษฎร์ธานี", "ภาคใต้"),
+    "85": ("ระนอง", "ภาคใต้"),
+    "86": ("ชุมพร", "ภาคใต้"),
+    "90": ("สงขลา", "ภาคใต้"),
+    "91": ("สตูล", "ภาคใต้"),
+    "92": ("ตรัง", "ภาคใต้"),
+    "93": ("พัทลุง", "ภาคใต้"),
+    "94": ("ปัตตานี", "ภาคใต้"),
+    "95": ("ยะลา", "ภาคใต้"),
+    "96": ("นราธิวาส", "ภาคใต้"),
+}
+
+
+def generate_thai_id() -> str:
+    digits = [random.randint(0, 9) for _ in range(12)]
+    checksum = sum(d * (13 - i) for i, d in enumerate(digits, start=1))
+    check_digit = (11 - (checksum % 11)) % 10
+    digits.append(check_digit)
+    return "".join(str(d) for d in digits)
+
+
+def get_province_info(citizen_id: str):
+    code = citizen_id[1:3]
+    return province_region_map.get(code, ("ไม่ทราบ", "ไม่ทราบ"))
+
+
+def show():
+    st.header("🎲 สุ่มเลขบัตรประชาชนไทย")
+    count = st.number_input("จำนวนเลขที่ต้องการสุ่ม", min_value=1, max_value=20, value=1, step=1)
+    if st.button("Generate", key="gen_thai_id"):
+        st.session_state.generated_ids = [generate_thai_id() for _ in range(count)]
+        for k in list(st.session_state.keys()):
+            if k.startswith("show_info_"):
+                del st.session_state[k]
+
+    ids = st.session_state.get("generated_ids", [])
+    for idx, cid in enumerate(ids):
+        col1, col2 = st.columns([3, 1])
+        col1.write(cid)
+        info_key = f"info_{idx}"
+        if col2.button("Information", key=info_key):
+            st.session_state[f"show_info_{idx}"] = not st.session_state.get(f"show_info_{idx}", False)
+        if st.session_state.get(f"show_info_{idx}", False):
+            prov, region = get_province_info(cid)
+            st.info(f"จังหวัด: {prov} | ภาค: {region}")
